Create future used paginating feature
Paginating will return a range of a bots message (if implemented) to
only be a set range. An example would be for commands like 'help' that
may return countless strings and then bot would spam the channel with a
large message. The use of this would make it so you could do "!help 1"
to return the first 'page' or range of command. The amount it returns is
configurable, but is set to a default of 6; However, can also be
overridden on a per function call basis.